Move get_instruction_encoding to VirtualMachine
